requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
PyMySQL==1.0.2
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
MarkupSafe==2.0.1
SQLAlchemy==1.4.46